refactor(frontend): enforce strict shared boundaries
