+++ frontend/src/components/my-events/calendarTagAccent.ts
@@ -1,5 +1,5 @@
 import type { EventItem } from "../../types/event";
-import { getEventFirstTagAccentClassNames } from "../../shared/tagAccent";
+import { getEventFirstTagAccentClassNames } from "../../features/events/lib/tagAccent";
 
 export const getCalendarEventAccentClassNames = (
   event: Pick<EventItem, "tags">,
