+++ frontend/src/pages/EventsPage.test.tsx
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import userEvent from "@testing-library/user-event";
 import { act, screen, waitFor, within } from "@testing-library/react";
-import { api } from "../shared/api";
+import { api } from "../shared/api/client";
 import { EventsPage } from "./EventsPage";
 import { fetchPublicEvents } from "../features/events/model/eventsSlice";
 import { useAssistantUiStore } from "../features/events/model/assistantUiStore";
@@ -620,4 +620,3 @@ describe("EventsPage", () => {
     expect(assistantInput).toHaveValue(suggestion);
   });
 });
-
